@@ -1,29 +1,30 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.2'
+    compileSdkVersion 29
 
     defaultConfig {
-        applicationId "org.traccar.client"
-        buildConfigField "boolean", "HIDDEN_APP", "false"
-        minSdkVersion 15
-        targetSdkVersion 28
-        versionCode 51
-        versionName '5.14'
+        applicationId 'org.traccar.client'
+        buildConfigField 'boolean', 'HIDDEN_APP', 'false'
+        minSdkVersion 16
+        targetSdkVersion 29
+        versionCode 62
+        versionName '6.1'
     }
 
     lintOptions {
         checkReleaseBuilds false
     }
 
-    flavorDimensions "default"
+    flavorDimensions 'default'
     productFlavors {
-        regular
+        regular {
+            isDefault.set(true)
+        }
         google
         hidden {
-            applicationId "org.traccar.client.hidden"
-            buildConfigField "boolean", "HIDDEN_APP", "true"
+            applicationId 'org.traccar.client.hidden'
+            buildConfigField 'boolean', 'HIDDEN_APP', 'true'
         }
     }
 
@@ -35,14 +36,16 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:design:28.0.0-rc02'
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.robolectric:robolectric:3.8'
-    googleImplementation 'com.google.firebase:firebase-core:16.0.3'
-    googleImplementation 'com.crashlytics.sdk.android:crashlytics:2.9.5'
+    implementation 'com.google.android.material:material:1.2.0-alpha04'
+    implementation 'androidx.preference:preference:1.1.0'
+    testImplementation 'junit:junit:4.13'
+    testImplementation 'org.robolectric:robolectric:4.1'
+    googleImplementation 'com.google.firebase:firebase-core:17.2.2'
+    googleImplementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
+    googleImplementation 'com.google.android.gms:play-services-location:17.0.0'
 }
 
-if (getGradle().getStartParameter().getTaskRequests().toString().contains("Google")) {
+if (getGradle().getStartParameter().getTaskRequests().toString().contains('Google')) {
     apply plugin: 'com.google.gms.google-services'
     apply plugin: 'io.fabric'
 
