@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 - 2017 Anton Tananaev (anton.tananaev@gmail.com)
+ * Copyright 2016 - 2019 Anton Tananaev (anton@traccar.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,28 +15,23 @@
  */
 package org.traccar.client;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.location.Criteria;
-import android.location.Location;
-import android.location.LocationListener;
-import android.location.LocationManager;
 import android.os.Bundle;
-import android.os.Looper;
 import android.preference.PreferenceManager;
-import android.support.annotation.DrawableRes;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.content.pm.ShortcutInfoCompat;
-import android.support.v4.content.pm.ShortcutManagerCompat;
-import android.support.v4.graphics.drawable.IconCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import androidx.annotation.DrawableRes;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.pm.ShortcutInfoCompat;
+import androidx.core.content.pm.ShortcutManagerCompat;
+import androidx.core.graphics.drawable.IconCompat;
 
 public class ShortcutActivity extends AppCompatActivity {
 
@@ -105,61 +100,30 @@ public class ShortcutActivity extends AppCompatActivity {
 
     @SuppressWarnings("MissingPermission")
     private void sendAlarm() {
-        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
-        LocationManager locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-
-        Criteria criteria = PositionProvider.getCriteria(
-                preferences.getString(MainFragment.KEY_ACCURACY, "medium"));
-
-        try {
-            Location location = locationManager.getLastKnownLocation(locationManager.getBestProvider(criteria, true));
-            if (location != null) {
-                sendAlarmLocation(location);
-            } else {
-                locationManager.requestSingleUpdate(criteria, new LocationListener() {
-                    @Override
-                    public void onLocationChanged(Location location) {
-                        sendAlarmLocation(location);
-                    }
-
-                    @Override
-                    public void onStatusChanged(String provider, int status, Bundle extras) {
-                    }
-
-                    @Override
-                    public void onProviderEnabled(String provider) {
-                    }
-
-                    @Override
-                    public void onProviderDisabled(String provider) {
-                    }
-                }, Looper.myLooper());
-            }
-        } catch (RuntimeException e) {
-            Toast.makeText(this, e.getMessage(), Toast.LENGTH_LONG).show();
-        }
-    }
-
-    private void sendAlarmLocation(Location location) {
-        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
-
-        Position position = new Position(
-                preferences.getString(MainFragment.KEY_DEVICE, null),
-                location, PositionProvider.getBatteryLevel(this));
-
-        String request = ProtocolFormatter.formatRequest(
-                preferences.getString(MainFragment.KEY_URL, null), position, ALARM_SOS);
-
-        RequestManager.sendRequestAsync(request, new RequestManager.RequestHandler() {
+        PositionProviderFactory.create(this, new PositionProvider.PositionListener() {
             @Override
-            public void onComplete(boolean success) {
-                if (success) {
-                    Toast.makeText(ShortcutActivity.this, R.string.status_send_success, Toast.LENGTH_SHORT).show();
-                } else {
-                    Toast.makeText(ShortcutActivity.this, R.string.status_send_fail, Toast.LENGTH_SHORT).show();
-                }
+            public void onPositionUpdate(Position position) {
+                SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(ShortcutActivity.this);
+                String request = ProtocolFormatter.formatRequest(
+                        preferences.getString(MainFragment.KEY_URL, null), position, ALARM_SOS);
+
+                RequestManager.sendRequestAsync(request, new RequestManager.RequestHandler() {
+                    @Override
+                    public void onComplete(boolean success) {
+                        if (success) {
+                            Toast.makeText(ShortcutActivity.this, R.string.status_send_success, Toast.LENGTH_SHORT).show();
+                        } else {
+                            Toast.makeText(ShortcutActivity.this, R.string.status_send_fail, Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
             }
-        });
+
+            @Override
+            public void onPositionError(Throwable error) {
+                Toast.makeText(ShortcutActivity.this, error.getMessage(), Toast.LENGTH_LONG).show();
+            }
+        }).requestSingleLocation();
     }
 
     private boolean executeAction(Intent intent) {
@@ -185,8 +149,7 @@ public class ShortcutActivity extends AppCompatActivity {
                     Toast.makeText(this, R.string.status_service_destroy, Toast.LENGTH_SHORT).show();
                     break;
                 case ACTION_SOS:
-                    if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED
-                            && ContextCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
+                    if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
                         sendAlarm();
                     } else {
                         Toast.makeText(this, R.string.status_send_fail, Toast.LENGTH_SHORT).show();
